Create button_pressed
Simple script for GameMaker: Studio. To be used in the Draw GUI event of any object. Creates a button from the sprite you name as the third argument. Use it as an if statement, so when clicked it'll run any code you place within that statement when the result is true (i.e. the button is clicked).

For example: 
if button_pressed(100, 100, spr_demo)
    {
    Run this code
    }